Create filtered-search
In the first searching I want to know who programs in Visual Studio. 
In the second searching I want to see it's historic of salary increase.
In the third, I want to see in which team it works.
In fourth, which projects it's team it's responsible for.
In the fifth searching, I want to know what tasks are from the teams responsabilities.